+ Test programs for control algorithms.

+++ test/EMAF_Test/EMAF_Test.c
@@ -6,9 +6,9 @@ int main ( void )
 {
     int i = 0;
 
-    double testArray[ 32 ] = { 1, 2, 10, 5, 15, 8, 12, 20, 10, 15, 16, 17, 18, 19, 20 };
-    double tempVal1 = 0;
-    double tempVal2 = 0;
+    float testArray[ 32 ] = { 1, 2, 10, 5, 15, 8, 12, 20, 10, 15, 16, 17, 18, 19, 20 };
+    float tempVal1 = 0;
+    float tempVal2 = 0;
 
     emaf_t test1;
     emaf_t test2;
